Move localization stuff into a separate file
